moved the parse fucntion out of lib.rs to fix wierd importing issues for it
